% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knitr_engine}
\alias{knitr_engine}
\title{Create a new engine for knitr}
\usage{
knitr_engine(envir)
}
\arguments{
\item{envir}{the environment in which llr code is executed}
}
\description{
Create a new engine for knitr
}
\examples{
\dontrun{
# in your first knitr chunk run
knitr::knit_engines$set(llr = llr::knitr_engine(sys.frame()))
}
}
