% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llr.R
\name{llr}
\alias{llr}
\title{Evaluates lisp-like R}
\usage{
llr(code, envir = parent.frame())
}
\arguments{
\item{code}{the code as string}

\item{envir}{the environment in which you want to run the code}
}
\description{
Evaluates lisp-like R
}
